+++ include/mach/atmega128/rtc.h
@@ -10,7 +10,7 @@ __BEGIN_SYS
 class ATMega128_RTC: public RTC_Common
 {
 private:
-    static const unsigned int EPOCH_DAYS = Traits<ATMega128_RTC>::EPOCH_DAYS;
+    static const unsigned long EPOCH_DAYS = Traits<ATMega128_RTC>::EPOCH_DAYS;
 
 public:
     ATMega128_RTC() {}
